+++ parser/src/test/java/io/github/cowwoc/styler/parser/test/ClassLiteralParserTest.java
@@ -18,6 +18,7 @@ import static io.github.cowwoc.styler.ast.core.NodeType.METHOD_INVOCATION;
 import static io.github.cowwoc.styler.ast.core.NodeType.PARAMETERIZED_TYPE;
 import static io.github.cowwoc.styler.ast.core.NodeType.QUALIFIED_NAME;
 import static io.github.cowwoc.styler.ast.core.NodeType.RETURN_STATEMENT;
+import static io.github.cowwoc.styler.ast.core.NodeType.WILDCARD_TYPE;
 import static io.github.cowwoc.styler.parser.test.ParserTestUtils.assertParseFails;
 import static io.github.cowwoc.styler.parser.test.ParserTestUtils.parseSemanticAst;
 import static io.github.cowwoc.styler.parser.test.ParserTestUtils.semanticNode;
@@ -51,6 +52,7 @@ public final class ClassLiteralParserTest
 			semanticNode(BLOCK, 24, 57),
 			semanticNode(PARAMETERIZED_TYPE, 28, 36),
 			semanticNode(QUALIFIED_NAME, 28, 33),
+			semanticNode(WILDCARD_TYPE, 34, 35),
 			semanticNode(CLASS_LITERAL, 41, 53),
 			semanticNode(IDENTIFIER, 41, 47));
 		requireThat(actual, "actual").isEqualTo(expected);
@@ -80,6 +82,7 @@ public final class ClassLiteralParserTest
 			semanticNode(BLOCK, 24, 67),
 			semanticNode(PARAMETERIZED_TYPE, 28, 36),
 			semanticNode(QUALIFIED_NAME, 28, 33),
+			semanticNode(WILDCARD_TYPE, 34, 35),
 			semanticNode(CLASS_LITERAL, 41, 63),
 			semanticNode(FIELD_ACCESS, 41, 50),
 			semanticNode(FIELD_ACCESS, 41, 57),
@@ -111,6 +114,7 @@ public final class ClassLiteralParserTest
 			semanticNode(BLOCK, 24, 62),
 			semanticNode(PARAMETERIZED_TYPE, 28, 36),
 			semanticNode(QUALIFIED_NAME, 28, 33),
+			semanticNode(WILDCARD_TYPE, 34, 35),
 			semanticNode(CLASS_LITERAL, 41, 58),
 			semanticNode(FIELD_ACCESS, 41, 52),
 			semanticNode(IDENTIFIER, 41, 46));
@@ -141,6 +145,7 @@ public final class ClassLiteralParserTest
 			semanticNode(BLOCK, 24, 54),
 			semanticNode(PARAMETERIZED_TYPE, 28, 36),
 			semanticNode(QUALIFIED_NAME, 28, 33),
+			semanticNode(WILDCARD_TYPE, 34, 35),
 			semanticNode(CLASS_LITERAL, 41, 50));
 		requireThat(actual, "actual").isEqualTo(expected);
 	}
@@ -169,6 +174,7 @@ public final class ClassLiteralParserTest
 			semanticNode(BLOCK, 24, 55),
 			semanticNode(PARAMETERIZED_TYPE, 28, 36),
 			semanticNode(QUALIFIED_NAME, 28, 33),
+			semanticNode(WILDCARD_TYPE, 34, 35),
 			semanticNode(CLASS_LITERAL, 41, 51));
 		requireThat(actual, "actual").isEqualTo(expected);
 	}
@@ -197,6 +203,7 @@ public final class ClassLiteralParserTest
 			semanticNode(BLOCK, 24, 59),
 			semanticNode(PARAMETERIZED_TYPE, 28, 36),
 			semanticNode(QUALIFIED_NAME, 28, 33),
+			semanticNode(WILDCARD_TYPE, 34, 35),
 			semanticNode(CLASS_LITERAL, 41, 55),
 			semanticNode(IDENTIFIER, 41, 47));
 		requireThat(actual, "actual").isEqualTo(expected);
@@ -226,6 +233,7 @@ public final class ClassLiteralParserTest
 			semanticNode(BLOCK, 24, 56),
 			semanticNode(PARAMETERIZED_TYPE, 28, 36),
 			semanticNode(QUALIFIED_NAME, 28, 33),
+			semanticNode(WILDCARD_TYPE, 34, 35),
 			semanticNode(CLASS_LITERAL, 41, 52));
 		requireThat(actual, "actual").isEqualTo(expected);
 	}
@@ -254,6 +262,7 @@ public final class ClassLiteralParserTest
 			semanticNode(BLOCK, 24, 58),
 			semanticNode(PARAMETERIZED_TYPE, 28, 36),
 			semanticNode(QUALIFIED_NAME, 28, 33),
+			semanticNode(WILDCARD_TYPE, 34, 35),
 			semanticNode(CLASS_LITERAL, 41, 54));
 		requireThat(actual, "actual").isEqualTo(expected);
 	}
@@ -309,6 +318,7 @@ public final class ClassLiteralParserTest
 			semanticNode(COMPILATION_UNIT, 0, 59),
 			semanticNode(CLASS_DECLARATION, 0, 58, "Test"),
 			semanticNode(METHOD_DECLARATION, 14, 56),
+			semanticNode(WILDCARD_TYPE, 20, 21),
 			semanticNode(BLOCK, 28, 56),
 			semanticNode(RETURN_STATEMENT, 32, 53),
 			semanticNode(CLASS_LITERAL, 39, 52),
@@ -340,6 +350,7 @@ public final class ClassLiteralParserTest
 			semanticNode(BLOCK, 24, 79),
 			semanticNode(PARAMETERIZED_TYPE, 28, 36),
 			semanticNode(QUALIFIED_NAME, 28, 33),
+			semanticNode(WILDCARD_TYPE, 34, 35),
 			semanticNode(CONDITIONAL_EXPRESSION, 41, 75),
 			semanticNode(IDENTIFIER, 41, 45),
 			semanticNode(CLASS_LITERAL, 48, 60),